requirements.txt
python-dotenv
flask
requests
jinja2